+++ src/adb/thirdparty/STFService/STFService.ts
@@ -10,6 +10,7 @@ import PromiseSocket from "promise-socket";
 import ThirdUtils from "../ThirdUtils";
 import * as STF from "./STFServiceModel";
 import STFProtoBuf from "./STFProtoBuf";
+import { Reader } from "protobufjs";
 
 const version = '2.4.9';
 
@@ -32,6 +33,10 @@ export interface STFServiceOptions {
    * local port use for STFService
    */
   agentPort: number,
+  /**
+   * calls timeout default is 15000 ms
+   */
+  timeout: number,  
 }
 
 const debug = Debug('STFService');
@@ -48,6 +53,7 @@ export default class STFService extends EventEmitter {
     this.config = {
       agentPort: 1090,
       servicePort: 1100,
+      timeout: 15000,
       ...config,
     }
   }
@@ -141,6 +147,7 @@ export default class STFService extends EventEmitter {
     void this.startAgentStream().catch((e) => { console.log('Agent failed', e); this.stop() });
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private handleResponse(id: number, result: any) {
     const resolv = this.responseHook[id];
     if (resolv) {
@@ -150,21 +157,33 @@ export default class STFService extends EventEmitter {
   }
 
   private async startServiceStream() {
-    let chunk: Buffer | null = null;
+    let buffer: Buffer | null = null;
     for (; ;) {
       await Utils.waitforReadable(this.servicesSocket);
       const next = await this.servicesSocket.read() as Buffer;
       if (!next)
         continue;
-      if (chunk) {
-        chunk = Buffer.concat([chunk, next]);
+      if (buffer) {
+        buffer = Buffer.concat([buffer, next]);
       } else {
-        chunk = next;
+        buffer = next;
       }
-      if (chunk) {
-        if (chunk.length == 1)
-          continue;
-        // console.log('servicesSocket RCV: ', chunk.length);
+      while (buffer) {
+        const reader = Reader.create(buffer);
+        
+        const envelopLen = reader.uint32();
+        const bufLen = envelopLen + reader.pos;
+
+        if (buffer.length < envelopLen)
+          break;
+        let chunk: Buffer;
+        if (bufLen === buffer.length) {
+          chunk = buffer;
+          buffer = null;
+        } else {
+          chunk = buffer.subarray(0, bufLen);
+          buffer = buffer.subarray(bufLen);
+        }
         try {
           const eventObj = this.proto.readEnvelope(chunk);
           const message = eventObj.message;
@@ -188,6 +207,7 @@ export default class STFService extends EventEmitter {
             case STF.MessageType.EVENT_BROWSER_PACKAGE:
               this.emit("browerPackage", this.proto.read.BrowserPackageEvent(message));
               break;
+            
             case STF.MessageType.GET_ACCOUNTS:
               this.handleResponse(eventObj.id, this.proto.read.GetAccountsResponse(message));
               break;
@@ -207,13 +227,31 @@ export default class STFService extends EventEmitter {
             case STF.MessageType.GET_PROPERTIES:
               this.handleResponse(eventObj.id, this.proto.read.GetBluetoothStatusResponse(message));
               break;
+            case STF.MessageType.GET_RINGER_MODE:
+              this.handleResponse(eventObj.id, this.proto.read.GetRingerModeResponse(message));
+              break;
+            case STF.MessageType.GET_ROOT_STATUS:
+              this.handleResponse(eventObj.id, this.proto.read.GetRootStatusResponse(message));
+              break;
+            case STF.MessageType.GET_SD_STATUS:
+              this.handleResponse(eventObj.id, this.proto.read.GetSdStatusResponse(message));
+              break;
+            case STF.MessageType.GET_VERSION:
+              this.handleResponse(eventObj.id, this.proto.read.GetVersionResponse(message));
+              break;
+            case STF.MessageType.GET_WIFI_STATUS:
+              this.handleResponse(eventObj.id, this.proto.read.GetWifiStatusResponse(message));
+              break;
 
+
+
+            
             default:
-              console.error('STFService not implemented Type:', eventObj.type);
+              console.error(`STFService Response Type (${eventObj.type}) is not implemented`);
           }
-          chunk = null;
+          // buffer = null;
         } catch (e) {
-          console.error(chunk.toString('hex'));
+          console.error(buffer.toString('hex'));
           console.error(e);
         }
       }
@@ -225,7 +263,6 @@ export default class STFService extends EventEmitter {
   private reqCnt = 1;
 
   private pushEnvelop<T>(envelope: STF.Envelope): Promise<T> {
-    // const id = Math.floor(Math.random() * 0xFFFFFF)
     const id = (this.reqCnt + 1) | 0xFFFFFF;
     this.reqCnt = id;
     envelope.id = this.reqCnt;
@@ -237,7 +274,7 @@ export default class STFService extends EventEmitter {
     const buf = this.proto.write.Envelope(envelope)
     this.servicesSocket.write(buf);
 
-    const timeout = Utils.delay(15000).then(() => {
+    const timeout = Utils.delay(this.config.timeout).then(() => {
       if (this.responseHook[id]) {
         delete this.responseHook[id];
         pReject(Error('timeout'));
@@ -252,92 +289,92 @@ export default class STFService extends EventEmitter {
     return this.pushEnvelop<STF.GetAccountsResponse>({ type: STF.MessageType.GET_ACCOUNTS, message })
   }
 
-  public async GetBrowsers(): Promise<STF.GetBrowsersResponse> {
+  public async getBrowsers(): Promise<STF.GetBrowsersResponse> {
     const message = this.proto.write.GetBrowsersRequest();
     return this.pushEnvelop<STF.GetBrowsersResponse>({ type: STF.MessageType.GET_BROWSERS, message })
   }
 
-  public async GetClipboard(type: STF.ClipboardType): Promise<STF.GetClipboardResponse> {
+  public async getClipboard(type: STF.ClipboardType): Promise<STF.GetClipboardResponse> {
     const message = this.proto.write.GetClipboardRequest({ type });
     return this.pushEnvelop<STF.GetClipboardResponse>({ type: STF.MessageType.GET_CLIPBOARD, message })
   }
 
-  public async GetDisplay(id: number): Promise<STF.GetDisplayResponse> {
+  public async getDisplay(id: number): Promise<STF.GetDisplayResponse> {
     const message = this.proto.write.GetDisplayRequest({ id });
     return this.pushEnvelop<STF.GetDisplayResponse>({ type: STF.MessageType.GET_DISPLAY, message })
   }
 
-  public async GetProperties(properties: string[]): Promise<STF.GetPropertiesResponse> {
+  public async getProperties(properties: string[]): Promise<STF.GetPropertiesResponse> {
     const message = this.proto.write.GetPropertiesRequest({ properties });
     return this.pushEnvelop<STF.GetPropertiesResponse>({ type: STF.MessageType.GET_PROPERTIES, message })
   }
 
-  public async GetRingerMode(): Promise<STF.GetRingerModeResponse> {
+  public async getRingerMode(): Promise<STF.GetRingerModeResponse> {
     const message = this.proto.write.GetRingerModeRequest();
     return this.pushEnvelop<STF.GetRingerModeResponse>({ type: STF.MessageType.GET_RINGER_MODE, message })
   }
 
-  public async GetSdStatus(): Promise<STF.GetSdStatusResponse> {
+  public async getSdStatus(): Promise<STF.GetSdStatusResponse> {
     const message = this.proto.write.GetSdStatusRequest();
     return this.pushEnvelop<STF.GetSdStatusResponse>({ type: STF.MessageType.GET_SD_STATUS, message })
   }
 
-  public async GetVersion(): Promise<STF.GetVersionResponse> {
+  public async getVersion(): Promise<STF.GetVersionResponse> {
     const message = this.proto.write.GetVersionRequest();
     return this.pushEnvelop<STF.GetVersionResponse>({ type: STF.MessageType.GET_VERSION, message })
   }
 
-  public async GetWifiStatus(): Promise<STF.GetWifiStatusResponse> {
+  public async getWifiStatus(): Promise<STF.GetWifiStatusResponse> {
     const message = this.proto.write.GetWifiStatusRequest();
     return this.pushEnvelop<STF.GetWifiStatusResponse>({ type: STF.MessageType.GET_WIFI_STATUS, message })
   }
 
-  public async GetBluetoothStatus(): Promise<STF.GetBluetoothStatusResponse> {
+  public async getBluetoothStatus(): Promise<STF.GetBluetoothStatusResponse> {
     const message = this.proto.write.GetBluetoothStatusRequest();
     return this.pushEnvelop<STF.GetBluetoothStatusResponse>({ type: STF.MessageType.GET_BLUETOOTH_STATUS, message })
   }
 
-  public async GetRootStatus(): Promise<STF.GetRootStatusResponse> {
+  public async getRootStatus(): Promise<STF.GetRootStatusResponse> {
     const message = this.proto.write.GetRootStatusRequest();
     return this.pushEnvelop<STF.GetRootStatusResponse>({ type: STF.MessageType.GET_ROOT_STATUS, message })
   }
 
-  public async SetClipboard(req: STF.SetClipboardRequest): Promise<STF.SetClipboardResponse> {
+  public async setClipboard(req: STF.SetClipboardRequest): Promise<STF.SetClipboardResponse> {
     const message = this.proto.write.SetClipboardRequest(req);
     return this.pushEnvelop<STF.SetClipboardResponse>({ type: STF.MessageType.SET_CLIPBOARD, message })
   }
 
-  public async SetKeyguardState(req: STF.SetKeyguardStateRequest): Promise<STF.SetKeyguardStateResponse> {
+  public async setKeyguardState(req: STF.SetKeyguardStateRequest): Promise<STF.SetKeyguardStateResponse> {
     const message = this.proto.write.SetKeyguardStateRequest(req);
     return this.pushEnvelop<STF.SetKeyguardStateResponse>({ type: STF.MessageType.SET_KEYGUARD_STATE, message })
   }
 
-  public async SetRingerMode(req: STF.SetRingerModeRequest): Promise<STF.SetRingerModeResponse> {
+  public async setRingerMode(req: STF.SetRingerModeRequest): Promise<STF.SetRingerModeResponse> {
     const message = this.proto.write.SetRingerModeRequest(req);
     return this.pushEnvelop<STF.SetRingerModeResponse>({ type: STF.MessageType.SET_RINGER_MODE, message })
   }
 
-  public async SetRotationRequest(req: STF.SetRotationRequest): Promise<any> {
+  public async setRotationRequest(req: STF.SetRotationRequest): Promise<any> {
     const message = this.proto.write.SetRotationRequest(req);
     return this.pushEnvelop<any>({ type: STF.MessageType.SET_ROTATION, message })
   }
 
-  public async SetWakeLock(req: STF.SetWakeLockRequest): Promise<STF.GetWifiStatusResponse> {
+  public async setWakeLock(req: STF.SetWakeLockRequest): Promise<STF.GetWifiStatusResponse> {
     const message = this.proto.write.SetWakeLockRequest(req);
     return this.pushEnvelop<STF.GetWifiStatusResponse>({ type: STF.MessageType.SET_WAKE_LOCK, message })
   }
 
-  public async SetWifiEnabledRequest(req: STF.SetWifiEnabledRequest): Promise<STF.SetWifiEnabledResponse> {
+  public async setWifiEnabledRequest(req: STF.SetWifiEnabledRequest): Promise<STF.SetWifiEnabledResponse> {
     const message = this.proto.write.SetWifiEnabledRequest(req);
     return this.pushEnvelop<STF.GetWifiStatusResponse>({ type: STF.MessageType.SET_WIFI_ENABLED, message })
   }
 
-  public async SetBluetoothEnabledRequest(req: STF.SetBluetoothEnabledRequest): Promise<STF.SetBluetoothEnabledResponse> {
+  public async setBluetoothEnabledRequest(req: STF.SetBluetoothEnabledRequest): Promise<STF.SetBluetoothEnabledResponse> {
     const message = this.proto.write.SetBluetoothEnabledRequest(req);
     return this.pushEnvelop<STF.SetBluetoothEnabledResponse>({ type: STF.MessageType.SET_BLUETOOTH_ENABLED, message })
   }
 
-  public async SetMasterMute(req: STF.SetMasterMuteRequest): Promise<STF.SetMasterMuteResponse> {
+  public async setMasterMute(req: STF.SetMasterMuteRequest): Promise<STF.SetMasterMuteResponse> {
     const message = this.proto.write.SetMasterMuteRequest(req);
     return this.pushEnvelop<STF.SetMasterMuteResponse>({ type: STF.MessageType.SET_MASTER_MUTE, message })
   }